fix(tests): locale es in .env.example and int casts in config/unblock
- .env.example APP_LOCALE: en → es (project is Spanish-first with en fallback;
  tests assert Spanish messages for ThrottleSimpleUnblock, FirewallException,
  CheckFirewallIntegration)
- config/unblock.php: cast env('REPORT_EXPIRATION') and env('HQ_WHITELIST_TTL')
  to int so Carbon::addSeconds() and HqWhitelistMail constructor (typed
  parameters under PHP 8.3 strict types) receive int when env vars come from
  .env.example shell values

Bug was only visible in an isolated test worktree (.env from .env.example,
APP_LOCALE=en); local dev had APP_LOCALE=es and did not define the TTL
envs, so config used its int defaults.

+++ config/unblock.php
@@ -5,7 +5,7 @@ return [
     'admin_email' => env('ADMIN_EMAIL'),
     'send_admin_report_email' => env('SEND_ADMIN_REPORT_EMAIL', true),
     'attempts' => env('ATTEMPTS', 10),
-    'report_expiration' => env('REPORT_EXPIRATION', 604800), // 7 days
+    'report_expiration' => (int) env('REPORT_EXPIRATION', 604800), // 7 days
     'cron_active' => env('CRON_ACTIVE', false),
 
     // Notification settings
@@ -67,7 +67,7 @@ return [
         'host_id' => env('HQ_HOST_ID'),
         'fqdn' => env('HQ_HOST_FQDN', ''),
         // Temporary whitelist TTL in seconds
-        'ttl' => env('HQ_WHITELIST_TTL', 14000),
+        'ttl' => (int) env('HQ_WHITELIST_TTL', 14000),
     ],
 
     // Simple Unblock Mode (v1.2.0+)
